+++ src/main/java/com/grookage/fsm/exceptions/InvalidStateException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.grookage.fsm.exceptions;
+package com.grookage.fsm.exceptions;
 
 /**
  * Entity by : koushikr.
