+++ src/main/java/com/grookage/fsm/models/entities/Context.java
@@ -13,16 +13,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.grookage.fsm.exceptions;
+package com.grookage.fsm.models.entities;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.io.Serializable;
 
 /**
  * Entity by : koushikr.
  * on 23/10/15.
+ *
+ * <p>
+ *     Denotes the Context in which the StateMachine will function
+ *     Keeps metadeta about the from and to states along with the causedEvent
+ * </p>
  */
-public class StateNotFoundException extends Exception {
+@Getter @Setter @AllArgsConstructor @NoArgsConstructor
+public class Context implements Serializable {
 
-    public StateNotFoundException(String errorMessage){
-        super(errorMessage);
-    }
+    private static final long serialVersionUID = 42L;
+
+    private State from;
+    private State to;
+    private Event causedEvent;
 
 }
