+++ src/main/java/com/grookage/fsm/models/entities/Transition.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.grookage.fsm.models.entities;
+package com.grookage.fsm.models.entities;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
